Revert "feat(condo): DOMA-8507 try to fix global mocking by jest"
This reverts commit 137d352c8355dc343ad3bb29d99da31d1d7d890d.
