+++ apps/condo/jest.config.js
@@ -9,8 +9,7 @@ module.exports = {
                 `${__dirname}/schema/**/*.test.js`,
                 `${__dirname}/domains/**/schema/*.test.js`,
             ],
-            setupFiles: [`${__dirname}/jest.setupTest.js`],
-            setupFilesAfterEnv: [`${__dirname}/jest.setupTestAfterEnv.js`],
+            setupFilesAfterEnv: [`${__dirname}/jest.setupTest.js`],
             // NOTE: need to pass uuid export syntax through babel
             transformIgnorePatterns: ['/node_modules/(?!(uuid|msgpackr)/)'],
         },
